[core] add deprecation module
