+++ ddtrace/contrib/elasticsearch/transport.py
@@ -1,17 +1,18 @@
 from elasticsearch import Transport
 from elasticsearch.exceptions import TransportError
 
-from .quantize import quantize
 from . import metadata
+from .quantize import quantize
+
+from ...utils.deprecation import deprecated
 from ...compat import urlencode
 from ...ext import AppTypes, http
-from ...util import deprecated
 
 DEFAULT_SERVICE = 'elasticsearch'
 SPAN_TYPE = 'elasticsearch'
 
 
-@deprecated(message='Use patching instead (see the docs).', version='0.6.0')
+@deprecated(message='Use patching instead (see the docs).', version='1.0.0')
 def get_traced_transport(datadog_tracer, datadog_service=DEFAULT_SERVICE):
 
     datadog_tracer.set_service_info(
