+++ ddtrace/contrib/mongoengine/patch.py
@@ -1,7 +1,7 @@
 import mongoengine
 
 from .trace import WrappedConnect
-from ddtrace.util import deprecated
+from ...utils.deprecation import deprecated
 
 # Original connect function
 _connect = mongoengine.connect
@@ -13,7 +13,7 @@ def patch():
 def unpatch():
     setattr(mongoengine, 'connect', _connect)
 
-@deprecated(message='Use patching instead (see the docs).', version='0.6.0')
+@deprecated(message='Use patching instead (see the docs).', version='1.0.0')
 def trace_mongoengine(*args, **kwargs):
     return _connect
