+++ ddtrace/contrib/pymysql/tracers.py
@@ -1,7 +1,8 @@
 import pymysql.connections
 
-from ddtrace.util import deprecated
+from ...utils.deprecation import deprecated
 
-@deprecated(message='Use patching instead (see the docs).', version='0.6.0')
+
+@deprecated(message='Use patching instead (see the docs).', version='1.0.0')
 def get_traced_pymysql_connection(*args, **kwargs):
     return pymysql.connections.Connection
