+++ ddtrace/contrib/sqlite3/connection.py
@@ -1,7 +1,8 @@
 from sqlite3 import Connection
 
-from ddtrace.util import deprecated
+from ...utils.deprecation import deprecated
 
-@deprecated(message='Use patching instead (see the docs).', version='0.6.0')
+
+@deprecated(message='Use patching instead (see the docs).', version='1.0.0')
 def connection_factory(*args, **kwargs):
     return Connection
